apps/dashboard/backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
httpx==0.27.2
python-multipart==0.0.9
PyJWT==2.10.1
supabase==2.27.0
supabase_auth==2.27.0
pytest==8.3.3
python-slugify==8.0.4
email-validator==2.2.0
openpyxl==3.1.5
xlrd==2.0.1
xlwt==1.3.0
openpyxl
xlrd
xlwt
